Defined "StrictCat", the category whose objects are all categories and whose morphisms are functors considered up to pointwise equality. Basic lemmas supporting the development added in Theory/Functor.v
Added a constructor for universal arrows.
Proved that, for a functor U : D -> C, if there is a universal arrow c -> U
for all c in C, then U has a left adjoint F.

Added a new constructor for adjunctions that does not make the user enter
redundant (dualized) data.

Did a bunch of work on the free category generated by a quiver in Free/Quiver.v.